refactor: some more reorganization
